+++ app/layout.tsx
@@ -1,10 +1,7 @@
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
-import { headers } from 'next/headers';
-import { cookieToInitialState } from '@account-kit/core';
 import './globals.css';
 import { Providers } from './providers';
-import { alchemyConfig } from '@/lib/alchemy/config';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -18,22 +15,12 @@ export const metadata: Metadata = {
   },
 };
 
-export default async function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  const initialState = cookieToInitialState(
-    alchemyConfig,
-    (await headers()).get('cookie') ?? undefined
-  );
-
+export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en">
       <body className={inter.className}>
-        <Providers initialState={initialState}>{children}</Providers>
+        <Providers>{children}</Providers>
       </body>
     </html>
   );
 }
-